+++ v0/destinations/iterable/transform.js
@@ -1,5 +1,6 @@
 const _ = require("lodash");
 const get = require("get-value");
+const { getCatalogEndpoint } = require("./util");
 const { EventType, MappedToDestinationKey } = require("../../../constants");
 const {
   ConfigCategory,
@@ -20,7 +21,8 @@ const {
   addExternalIdToTraits,
   isAppleFamily,
   handleRtTfSingleEventError,
-  checkInvalidRtTfEvents
+  checkInvalidRtTfEvents,
+  getDestinationExternalIDInfoForRetl
 } = require("../../util");
 const logger = require("../../../logger");
 
@@ -234,6 +236,9 @@ function constructPayloadItem(message, category, destination) {
     case "alias":
       rawPayload = constructPayload(message, mappingConfig[category.name]);
       break;
+    case "catalogs":
+      rawPayload = constructPayload(message, mappingConfig[category.name]);
+      break;
     default:
       logger.debug("not supported type");
   }
@@ -243,9 +248,16 @@ function constructPayloadItem(message, category, destination) {
 
 function responseBuilderSimple(message, category, destination) {
   const response = defaultRequestConfig();
-  response.endpoint = category.endpoint;
+  response.endpoint =
+    category.action === "catalogs"
+      ? getCatalogEndpoint(category, message)
+      : category.endpoint;
   response.method = defaultPostRequestConfig.requestMethod;
   response.body.JSON = constructPayloadItem(message, category, destination);
+  // adding operation to understand what type of event will be sent in batch
+  if (category.action === "catalogs") {
+    response.operation = "catalogs";
+  }
   response.headers = {
     "Content-Type": "application/json",
     api_key: destination.Config.apiKey
@@ -286,7 +298,17 @@ function processSingleMessage(message, destination) {
   let event;
   switch (messageType) {
     case EventType.IDENTIFY:
-      category = ConfigCategory.IDENTIFY;
+      if (
+        get(message, MappedToDestinationKey) &&
+        getDestinationExternalIDInfoForRetl(message, "ITERABLE").objectType !==
+          "users"
+      ) {
+        // catagory will be catalog for any other object other than users
+        // DOC: https://support.iterable.com/hc/en-us/articles/360033214932-Catalog-Overview
+        category = ConfigCategory.CATALOG;
+      } else {
+        category = ConfigCategory.IDENTIFY;
+      }
       break;
     case EventType.PAGE:
       category = ConfigCategory.PAGE;
@@ -342,6 +364,11 @@ function batchEvents(arrayChunks) {
   arrayChunks.forEach(chunk => {
     const batchResponseList = [];
     const metadatas = [];
+    // DOC: https://api.iterable.com/api/docs#catalogs_bulkUpdateCatalogItems
+    const batchCatalogResponseList = {
+      documents: {},
+      replaceUploadedFieldsOnly: true
+    };
 
     // extracting destination
     // from the first event in a batch
@@ -352,11 +379,34 @@ function batchEvents(arrayChunks) {
 
     // Batch event into dest batch structure
     chunk.forEach(ev => {
+      if (chunk[0].message.operation === "catalogs") {
+        // body will be in the format:
+        //   {
+        //     "documents": {
+        //         "test-1-item": {
+        //             "abc": "TestValue"
+        //         },
+        //         "test-2-item": {
+        //             "abc": "TestValue1"
+        //         }
+        //     },
+        //     "replaceUploadedFieldsOnly": true
+        // }
+        batchCatalogResponseList.documents[
+          ev.message.endpoint.split("/").pop()
+        ] = get(ev, "message.body.JSON.update");
+      }
       batchResponseList.push(get(ev, "message.body.JSON"));
       metadatas.push(ev.metadata);
     });
     // batching into identify batch structure
-    if (chunk[0].message.endpoint.includes("/api/users")) {
+    if (chunk[0].message.operation === "catalogs") {
+      batchEventResponse.batchedRequest.body.JSON = batchCatalogResponseList;
+      batchEventResponse.batchedRequest.endpoint = chunk[0].message.endpoint.substr(
+        0,
+        chunk[0].message.endpoint.lastIndexOf("/")
+      );
+    } else if (chunk[0].message.endpoint.includes("/api/users")) {
       batchEventResponse.batchedRequest.body.JSON = {
         users: batchResponseList
       };
@@ -400,7 +450,10 @@ function getEventChunks(
 ) {
   // Categorizing identify and track type of events
   // Checking if it is identify type event
-  if (event.message.endpoint.includes("api/users/update")) {
+  if (
+    event.message.endpoint.includes("api/users/update") ||
+    event.message.operation === "catalogs"
+  ) {
     identifyEventChunks.push(event);
   } else if (event.message.endpoint.includes("api/events/track")) {
     // Checking if it is track type of event